Initiative: BS5 updates; badge init counts - add support for abstracted JS methods
