+++ app/views/combatants/index.html.erb
@@ -1,11 +1,12 @@
 <% @combatants.each do |c| %>
-  <% classes = ['list-group-item'] %>
+  <% classes = %w(list-group-item d-flex justify-content-between align-items-center) %>
   <% classes << 'list-group-item-info' if c.active %>
   <%= tag.div(class: classes) do %>
-    <%= tag.span(c.id, class: %w(init-id d-none), aria: { hidden: true }) %>
-    <%= tag.a('', class: %w(fas fa-edit edit-btn)) %>
-    <%= tag.a('', class: %w(fas fa-trash delete-btn)) %>
-    <%= tag.span(c.time, class: %w(init-time)) %>
-    <%= tag.span(c.name, class: %w(init-name)) %>
+    <%= tag.div(class: 'object-id', 'data-object-id': c.id) do %>
+      <%= tag.i(class: %w(bi bi-pencil-fill text-primary edit-button)) %>
+      <%= tag.i(class: %w(bi bi-trash-fill text-primary delete-button)) %>
+      <%= tag.span(c.name, class: %w(ms-3)) %>
+    <% end %>
+    <%= tag.span(c.time, class: %w(badge bg-primary rounded-pill)) %>
   <% end %>
 <% end %>
