+++ app/webpacker/packs/caster_classes.js
@@ -1,5 +1,6 @@
 import {Helpers} from "../src/javascript/ajax_helpers";
 import {Modal} from "bootstrap";
+import {AbstractMethods} from "../src/javascript/abstract_methods";
 
   /* Add validation code here;
      See https://getbootstrap.com/docs/4.5/components/forms/#validation
@@ -51,20 +52,7 @@ function openNewModal() {
     .then(prepareEditForm);
 }
 
-function prepareEditForm(ajaxBody) {
-  const myModal = document.getElementById('object-modal');
-  myModal.innerHTML = ajaxBody;
-
-  const myForm = document.getElementById('object-form');
-  const submitButton = document.getElementById('object-modal-ok');
-  submitButton.addEventListener('click', _ =>
-    Helpers.submitFormAndReloadPage(myForm, fetchResults));
-
-  const formInputs = myForm.querySelectorAll('input.form-control');
-  myModal.addEventListener('shown.bs.modal', _ => { formInputs[0].focus() });
-
-  new Modal(myModal).show();
-}
+const prepareEditForm = AbstractMethods.prepareEditForm(fetchResults);
 
 window.addEventListener('load', () => {
   fetchResults();
