+++ app/webpacker/src/stylesheets/application.scss
@@ -24,29 +24,11 @@
   }
 }
 
-a.fas {
-  text-decoration: none;
-}
-
 span.edit-confirm {
   margin-left: 20px;
 
   span.btn {
     margin-left: 10px;
-  }
-}
-
-.edit-icon {
-  margin-left: 25px;
-  opacity: 0.4;
-}
-
-#initiative-list {
-  .init-time {
-    display: inline-block;
-    width: 2em;
-    margin-left: 20px;
-    margin-right: 10px;
   }
 }
